some minor fixes for benchmark

+++ src/test/app/PerformanceCounters.cpp
@@ -168,7 +168,7 @@ void PerformanceCounters::fetch() {
     mTimeEnabledNanos = mReadFormat[1];
     mTimeRunningNanos = mReadFormat[2];
 
-    for (uint64_t i = 0; i < mReadFormat[0]; i++) {
+    for (size_t i = 0; i < mReadFormat[0]; i++) {
         auto val = mReadFormat[3 + i * 2 + 0];
         auto id = mReadFormat[3 + i * 2 + 1];
         auto it = mIdToValue.find(id);
